refactor Filter into FilterSelect and MileageInput, move CustomDropdownIndicator, customStyles, format functions, preserve styles and functionality
