+++ index.html
@@ -2,19 +2,13 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link
-      rel="icon"
-      type="image/png"
-      href="/public/favicon.png"
-      sizes="32x32"
-    />
+    <link rel="icon" type="image/png" href="/favicon.png" sizes="32x32" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <meta
       name="description"
       content="RentalCar - Find your perfect rental car for any journey"
     />
     <title>Car Rental</title>
-    <!-- Додаємо шрифт Manrope через Google Fonts -->
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link
