+++ src/pages/CatalogPage/CatalogPage.module.css
@@ -9,7 +9,7 @@
 .grid {
   display: grid;
   grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
-  gap: 48px;
+  row-gap: 48px;
   justify-items: center;
   margin-top: 2rem;
 }
